ctype macros should take explicitly unsigned input
to avoid chars with high bits being converted to negative numbers

perl -p -i -e 's/(tolower|toupper|isupper|islower|isdigit|isalpha|isspace|isxdigit)\(\s*\*/${1}((uint8_t) */g' $(find . -name "*.[ch]" -print)
perl -p -i -e 's/(tolower|toupper|isupper|islower|isdigit|isalpha|isspace|isxdigit)\(\(int\)/${1}((uint8_t)/g' $(find . -name "*.[ch]" -print)

+++ src/lib/getaddrinfo.c
@@ -287,7 +287,7 @@ int getaddrinfo(char const *hostname, char const *servname, struct addrinfo cons
 	}
 
 	if (servname) {
-		if (isdigit((int)*servname)) {
+		if (isdigit((uint8_t)*servname)) {
 			port = htons(atoi(servname));
 		} else {
 			struct		servent *se;
